Fix (again!) wrong definition of length in extended blocks
The length of a block element should be just that of the same element in the
parent block. What could be simpler? Nonetheless, a wrong definition, counting
the ascents in the extended block instead, was not only implemented initially,
but then after having been corrected in |extended_block| on 13 Jul 2013
(commit 549ea34), was independently reintroduced on 9 Sep 2015 (commit 3e3ac8)
in the very similar |ext_block| class that was to eventually replace the
|extended_block|. I just hope I won't reintroduce this error a third time.